+++ tasmota/include/i18n.h
@@ -416,6 +416,7 @@
 #define D_CMND_SENSORRETAIN "SensorRetain"
 #define D_CMND_INFORETAIN "InfoRetain"
 #define D_CMND_STATERETAIN "StateRetain"
+#define D_CMND_STATUSRETAIN "StatusRetain"
 #define D_CMND_PUBLISH "Publish"
 #define D_CMND_FILEUPLOAD "FileUpload"
 #define D_CMND_FILEDOWNLOAD "FileDownload"